+++ src/components/burger-constructor/burger-constructor.module.css
@@ -9,11 +9,10 @@
     display: flex;
     flex-direction: column;
     gap: 16px;
-    height: calc(100vh - 415px);
+    height: max(calc(100vh - 415px), 390px);
 }
 
 .unlockedComponents {
-    /*composes: componentsList;*/
     display: flex;
     flex-direction: column;
     gap: 16px;
@@ -30,7 +29,7 @@
 .order {
     display: flex;
     justify-content: flex-end;
-    align-self: end;
+    /*align-self: end;*/
 }
 
 .total {
