+++ src/components/burger-ingredients/burger-ingredients.jsx
@@ -1,13 +1,13 @@
 import React, {useEffect, useRef} from "react";
 import {useSelector, useDispatch} from "react-redux";
-import {fetchIngredients} from "../../services/ingredientsSlice";
+import {fetchIngredients} from "../../services/burger-ingredients-slice";
 
 import styles from "./burger-ingredients.module.css"
-import Tabs from "../tabs/tabs"
-import IngredientsCategory from "../ingredients-category/ingredients-category";
+import Tabs from "./tabs/tabs"
+import IngredientsCategory from "./ingredients-category/ingredients-category";
 import Modal from "../modal/modal";
-import IngredientDetails from "../ingredient-details/ingredient-details";
-import {ingredientDetailsClosed} from "../../services/ingredientDetailsSlice";
+import IngredientDetails from "./ingredient-details/ingredient-details";
+import {ingredientDetailsClosed} from "../../services/ingredient-details-slice";
 
 
 export default function BurgerIngredients() {
