+++ src/components/burger-ingredients/ingredients-category/ingredients-category.jsx
@@ -3,9 +3,9 @@ import {useDispatch} from "react-redux";
 import PropTypes from "prop-types";
 
 import styles from "./ingredients-category.module.css"
-import {ingredientPropType} from "../../utils/prop-types"
+import {ingredientPropType} from "../../../utils/prop-types"
 import Card from "../card/card";
-import {activeTabSelected} from "../../services/ingredientsSlice";
+import {activeTabSelected} from "../../../services/burger-ingredients-slice";
 
 export default function IngredientsCategory({ingredients, categoryName, className}) {
