--- src/components/dnd-mockup/dnd-mockup.module.css
@@ -1,36 +1,0 @@
-.section {
-    width: 600px;
-}
-
-.componentsList {
-    list-style: none;
-    padding: 0;
-    margin: 0;
-    display: flex;
-    flex-direction: column;
-    gap: 16px;
-    height: calc(100vh - 415px);
-}
-
-.unlockedComponents {
-    composes: componentsList;
-    height: auto;
-    overflow-y: scroll;
-}
-
-.component {
-    display: flex;
-    align-items: center;
-    gap: 8px;
-}
-
-.order {
-    display: flex;
-    justify-content: flex-end;
-    align-self: end;
-}
-
-.total {
-    display: flex;
-    align-items: center;
-}
