+++ src/components/header/header.jsx
@@ -1,6 +1,6 @@
 import styles from "./header.module.css"
 import { Logo, BurgerIcon, ListIcon, ProfileIcon } from "@ya.praktikum/react-developer-burger-ui-components";
-import NavigationItem from "../navigation-item/navigation-item";
+import NavigationItem from "./navigation-item/navigation-item";
 
 export default function AppHeader() {
     return (
